requirements.txt
academictorrents==2.3.3
accelerate==0.27.2
addict==2.4.0
aiofiles==24.1.0
aiohttp==3.9.5
anyio==4.4.0
apex==0.9.10dev
azureml==0.2.7
basicsr==1.4.2
blend_modes==2.1.0
blind_watermark==0.4.4
cachetools==5.3.3
caffe2==0.8.1
cityscapesscripts==2.2.3
clip_interrogator==0.6.0
color_matcher==0.5.0
color_matcher
colour==0.1.5
compel==2.0.2
cuml==0.6.1.post1
cupy_cuda12x==12.3.0
easyocr==1.7.1
einops==0.8.0
external==0.0.1
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.111.0
fsspec==2024.3.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gfpgan==1.3.8
GitPython==3.1.43
GitPython==3.1.43
huggingface_hub==0.23.1
hydra-core==1.3.2
image_reward==1.5
imageio==2.34.1
imageio_ffmpeg==0.4.9
img2texture==1.0.6
importlib_metadata==7.1.0
intel_extension_for_pytorch==2.3.100
iopath==0.1.10
ipdb==0.13.13
joblib==1.4.2
kornia==0.7.2
lark==1.1.9
lark_parser==0.12.0
librosa==0.10.2.post1
lmdb==1.5.1
loguru==0.7.2
mano==0.5.2
matplotlib==3.9.1
matrix_client==0.4.0
mc==0.0
mediapipe==0.10.14
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.4
more_itertools==10.3.0
mysql_connector_repackaged==0.3.1
numba==0.59.1
numexpr==2.10.0
numpy==2.0.0
nvidia_smi==0.1.3
omegaconf==2.3.0
onnxruntime==1.18.0
onnxruntime_gpu==1.18.0
open3d==0.18.0
open_clip_torch==2.24.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.82
opendr==0.78
openvino==2024.2.0
packaging==24.1
pandas==2.2.2
parrots==1.0.3
piexif==1.1.3
pilgram==1.2.1
Pillow==10.4.0
pip==24.0
pixeloe==0.0.10
plyfile==1.0.3
prettytable==3.10.0
protobuf==5.27.2
psd_tools==1.9.33
psutil==5.9.0
psutil==5.9.8
pycocotools==2.0.7
pydantic==2.8.2
pygit2==1.15.0
Pygments==2.18.0
PyMatting==1.1.12
pytest==8.2.2
python-dotenv==1.0.1
python-socketio==5.11.3
python_magic==0.4.27
pytorch_fid==0.3.0
pytorch_lightning==2.2.5
pyunpack==0.3
PyYAML==6.0.1
PyYAML==6.0.1
pyzbar==0.1.9
regex==2024.5.15
Requests==2.32.3
requests_toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scikit_learn==1.5.1
scipy==1.14.0
setuptools==69.5.1
Shapely==2.0.4
skimage==0.0
spandrel==0.3.4
spandrel_extra_arches==0.1.1
starlette==0.37.2
submitit==1.5.1
supervision==0.21.0
tabulate==0.9.0
taichi==1.7.1
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow_datasets==4.9.6
termcolor==2.4.0
tifffile==2024.5.22
timm==0.6.13
torch==2.3.0+cu121
torch_directml==0.2.2.dev240614
torch_tb_profiler==0.4.3
torchaudio==2.3.0+cu121
torchmetrics==0.11.4
torchsde==0.2.6
torchvision==0.18.0+cu121
tqdm==4.66.4
transformers==4.38.2
trimesh==4.4.0
turbojpeg==0.0.2
typer==0.12.3
typer_config==1.4.0
typing_extensions==4.12.2
uvicorn==0.30.1
vapoursynth==69
websocket_client==1.8.0
wget==3.2
xformers==0.0.26.post1
yacs==0.1.8
yapf==0.40.2
